fix: update wallet identifier for Binance and restructure token definitions for consistency
